refactor: flatten architecture from features/shared to components/lib/types
Move from feature-based organization (src/features/, src/shared/) to a
flat structure (src/components/, src/lib/, src/types/) following Vercel
Commerce conventions. Rename files to kebab-case and remove barrel exports.
