+++ src/components/layout/user-menu.tsx
@@ -1,7 +1,7 @@
-import { auth } from '@/features/auth/lib/auth';
-import { UserMenuPresenter } from './UserMenuPresenter';
+import { auth } from '@/lib/auth';
+import { UserMenuPresenter } from './user-menu-presenter';
 
-export async function UserMenuContainer() {
+export async function UserMenu() {
   const session = await auth();
   const name = session?.user?.name;
   const email = session?.user?.email;
