+++ src/components/ui/avatar.tsx
@@ -3,7 +3,7 @@
 import * as AvatarPrimitive from '@radix-ui/react-avatar';
 import * as React from 'react';
 
-import { cn } from '@/shared/lib/utils';
+import { cn } from '@/lib/utils';
 
 function Avatar({
   className,
