--- src/features/auth/actions/sign-out.action.ts
@@ -1,7 +1,0 @@
-'use server';
-
-import { signOut } from '@/features/auth/lib/auth';
-
-export async function signOutAction() {
-  await signOut();
-}
