--- src/features/auth/components/server/AuthForm/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthForm';
