--- src/features/auth/components/server/AuthPageTemplate/index.ts
@@ -1,1 +1,0 @@
-export * from './AuthPageTemplate';
