--- src/features/auth/components/server/ProviderIcon/index.ts
@@ -1,1 +1,0 @@
-export * from './ProviderIcon';
