--- src/shared/components/layout/server/Header/UserMenu/UserMenu.tsx
@@ -1,1 +1,0 @@
-export { UserMenuContainer as UserMenu } from './UserMenuContainer';
