--- src/shared/components/layout/server/Header/UserMenu/index.ts
@@ -1,1 +1,0 @@
-export { UserMenu } from './UserMenu';
